DS_Compliance_OCR_Data_Extractor-master/backend/pyproject.toml
[project]
name = "Oryx"
version = "0.1.0"
description = "Oryx - Document OCR Engine"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "Django>=5.0,<6",
    "django-environ>=0.11",
    "psycopg[binary]>=3.1",
    "celery[redis]>=5.3",
    "redis>=5.0",
    "opencv-python-headless>=4.9",
    "numpy>=1.26",
    "Pillow>=10.2",
    "PyMuPDF>=1.23",
    "pypdfium2>=4.28",
    "azure-ai-documentintelligence>=1.0",
    "azure-core>=1.30",
    "pandas>=2.0",
    "openpyxl>=3.1",
    "python-docx>=1.1",
    "httpx>=0.27",
    "gunicorn>=21.2",
    "cryptography>=42.0",
    "pymongo>=4.6",
    "djangosaml2>=1.9",
    "pyopenssl>=24.0.0",
    "psycopg2-binary>=2.9.11",
]
